curd模版

+++ oa/src/main/java/com/lubuwei2/ssm/domain/FindMapper.java
@@ -3,7 +3,7 @@ package com.lubuwei2.ssm.domain;
 import java.util.List;
 
 public interface FindMapper<T, D> {
-    D findByPK(Integer id); // 根据主键查找
+    D findByPK(Long id); // 根据主键查找
 
     List<D> find(T domain);
 }
